requirements.txt
beautifulsoup4==4.12.2
keybert==0.7.0
keyboard==0.13.5
newspaper==0.1.0.7
newspaper3k==0.2.8
progressbar33==2.4
requests==2.28.2
summa==1.2.0
termcolor==1.1.0